+++ src/modules/index.js
@@ -4,3 +4,5 @@ exports.date = require('./date').date
 exports.ddmmyyyy = require('./date').ddmmyyyy
 exports.mmyyyy = require('./date').mmyyyy
 exports.yyyy = require('./date').yyyy
+
+exports.makeId = require('./makeId')