+++ src/redux/actions/sale.js
@@ -1,0 +1,15 @@
+'use strict'
+
+export function localStorageSale(data) {
+	return {
+		type: 'LOCAL_STORAGE_SALE',
+		data
+	}
+}
+
+export function penjualan(data) {
+	return {
+		type: 'PENJUALAN',
+		data
+	}
+}