+++ src/redux/reducers/index.js
@@ -4,10 +4,20 @@ import { combineReducers } from 'redux'
 
 import user from './user'
 import category from './category'
+import sale from './sale'
 
-const rootReducer = combineReducers({
+const appReducer = combineReducers({
 	user,
-	category
+	category,
+	sale
 })
+
+const rootReducer = (state, action) => {
+	if(action.type === 'LOGOUT') {
+		state = undefined
+	}
+
+	return appReducer(state, action)
+}
 
 export default rootReducer