+++ src/redux/reducers/sale.js
@@ -1,23 +1,54 @@
 'use strict'
 
+import {
+	AsyncStorage
+} from 'react-native'
+
 /*
 *
 sale
 *
 */
 /*
-{
-	idSale: string,
-	amount: number,
-	discount: number,
-	total: number,
-	cash: number,
-	change: number,
-	data: [{
-		idProduct: string,
+data: [
+	{
+		idSale: string,
 		amount: number,
 		discount: number,
-		toal: number
-	}]
-}
+		total: number,
+		cash: number,
+		change: number,
+		data: [{
+			idProduct: string,
+			amount: number,
+			discount: number,
+			total: number
+		}]
+	}
+]
 */
+
+const initialState = {
+	data: []
+}
+
+const SaleReducers = (state = initialState, action) => {
+	switch(action.type) {
+		case 'LOCAL_STORAGE_SALE':
+			return {
+				...state,
+				data: action.data
+			}
+
+		case 'PENJUALAN':
+			AsyncStorage.setItem('@Penjualan', JSON.stringify([...state.data, action.data]))
+			return {
+				data: [...state.data, action.data]
+			}
+
+		default:
+			return state
+	}
+}
+
+export default SaleReducers